spreadSheetのリゾルバを作成
