+++ server/src/graphql/resolvers/index.ts
@@ -8,6 +8,7 @@ import { TechBranch, TechLeaf, TechTree } from "../../models/TechForest.model";
 import { UserLeafs, UserUrls } from "../../models/User.model";
 import { IdType } from "../../types";
 import { specSheetMutations, specSheetQueries } from "./specSheet";
+import { sepreadSheetQueries, spreadSheetMutations } from "./spreadSheet";
 import { studyStackMutations, studyStackQueries } from "./studyStack";
 import { techForestMutations, techForestQueries } from "./techForest";
 import { userMutations, userQueries } from "./user";
@@ -22,6 +23,7 @@ const resolvers = {
     ...techForestQueries,
     ...studyStackQueries,
     ...specSheetQueries,
+    ...sepreadSheetQueries,
   },
   Mutation: {
     ...userMutations,
@@ -31,6 +33,7 @@ const resolvers = {
     ...techForestMutations,
     ...studyStackMutations,
     ...specSheetMutations,
+    ...spreadSheetMutations,
   },
   TechBranch: {
     techLeafs: async ({ _id }: IdType) =>
