+++ server/src/graphql/resolvers/spreadSheet/index.ts
@@ -1,0 +1,2 @@
+export { default as spreadSheetMutations } from "./mutations";
+export { default as sepreadSheetQueries } from "./queries";
