+++ server/src/graphql/resolvers/spreadSheet/mutations.ts
@@ -1,0 +1,7 @@
+import { UserIdType } from "../../../types";
+
+const spreadSheetMutations = {
+  // updateSpreadSheet: (_: any, { userId }: UserIdType) => {},
+};
+
+export default spreadSheetMutations;
